requirements.txt
openai-whisper==20231117
torch>=2.2.0
pydub==0.25.1
srt==3.5.3
python-dotenv==1.0.0
moviepy==1.0.3 